Update dscg_controls.md
A number of minor changes: Some punctuation; The omission of a superfluous "the;" brevity, in the form of "The... DSCG refer to" becoming "is;" (all naming is referential. Calling a thing what it is uses the word "is" almost always in English)

RE: "predicates," I'm suggesting "conditions," but "limits" might also be the case here. "Predicates" is unnecessarily Latinate and complex, especially for non-native speakers. "Conditions" or "limits" covers it.